+++ CHANGELOG.md
@@ -763,7 +763,8 @@ This major version introduces the concept of `Engines` which handle the chunking
 
 - Inputs:
   - Renamed `FieldDataNC` as `FieldData`, now with improved data handling
-  - New states class `WeibullPointCloud`, representing Weibull sector data with support at arbitrary point locations
+  - New states class `WeibullField`, representing Weibull sector data on a regular grid
+  - New states class `WeibullPointCloud`, representing Weibull sector data at arbitrary point locations
   - Updated `windio` inputs
 - Models:
   - New turbine type `DirectDataType`, for direct infusion of farm data at a turbine